+++ analyser/rule.go
